requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
cloudinary
werkzeug
Jinja2
gunicorn
